# nodejs-server-server/api/swagger.yaml
...
swagger: "2.0"
info:
  description: "This is Swagger UI environment generated for the TMF Service Catalog\
    \ Management specification"
  version: "2.0.0"
  title: "Service Catalog Management"
host: "localhost:8080"
basePath: "/tmf-api/serviceCatalogManagement/v2/"
schemes:
- "https"
consumes:
- "application/json"
produces:
- "application/json"
paths:
...
  /serviceSpecification:
    get:
      tags:
      - "serviceSpecification"
      summary: "List or find 'ServiceSpecification' objects"
      operationId: "listServiceSpecification"
      parameters:
      - name: "fields"
        in: "query"
        description: "Comma separated properties to display in response"
        required: false
        type: "string"
      - name: "name"
        in: "query"
        description: "For filtering: Name given to this REST resource"
        required: false
        type: "string"
      - name: "type"
        in: "query"
        description: "For filtering: Class type of this REST resource"
        required: false
        type: "string"
      - name: "@schemaLocation"
        in: "query"
        description: "For filtering: This field provides a link to the schema describing\
          \ this REST resource"
        required: false
        type: "string"
      - name: "@baseType"
        in: "query"
        description: "For filtering: The (immediate) base class type of this REST\
          \ resource"
        required: false
        type: "string"
      - name: "version"
        in: "query"
        description: "For filtering: Service specification version"
        required: false
        type: "string"
      - name: "validForStartDateTime"
        in: "query"
        description: "For filtering: An instant of time, starting at the TimePeriod"
        required: false
        type: "string"
        format: "date-time"
      - name: "validForEndDateTime"
        in: "query"
        description: "For filtering: An instant of time, ending at the TimePeriod."
        required: false
        type: "string"
        format: "date-time"
      - name: "lastUpdate"
        in: "query"
        description: "For filtering: Date and time of the last update of this REST\
          \ resource"
        required: false
        type: "string"
        format: "date-time"
      - name: "lifecycleStatus"
        in: "query"
        description: "For filtering: Used to indicate the current lifecycle status\
          \ of the service specification"
        required: false
        type: "string"
      - name: "isBundle"
        in: "query"
        description: "For filtering: A flag indicates that if this service specification\
          \ is a bundled specification (true) or single (false)."
        required: false
        type: "boolean"
      - name: "resourceSpecificationVersion"
        in: "query"
        description: "For filtering: ResourceSpecification version"
        required: false
        type: "string"
      - name: "resourceSpecificationName"
        in: "query"
        description: "For filtering: Name given to the ResourceSpecification"
        required: false
        type: "string"
      - name: "resourceSpecificationType"
        in: "query"
        description: "For filtering: The (class) type of the ResourceSpecification"
        required: false
        type: "string"
      - name: "attachmentName"
        in: "query"
        description: "For filtering: name given to the attachment"
        required: false
        type: "string"
      - name: "attachmentType"
        in: "query"
        description: "For filtering: Attachment class  type"
        required: false
        type: "string"
      - name: "attachmentUri"
        in: "query"
        description: "For filtering: Uniform Resource Identifier (URI) of the attachment"
        required: false
        type: "string"
      - name: "attachmentMimeType"
        in: "query"
        description: "For filtering: Attachment mime type such as extension file for\
          \ video, picture and document"
        required: false
        type: "string"
      - name: "serviceSpecCharacteristicName"
        in: "query"
        description: "For filtering: A word, term, or phrase by which this characteristic\
          \ specification is known and distinguished from other characteristic specifications."
        required: false
        type: "string"
      - name: "serviceSpecCharacteristicValueType"
        in: "query"
        description: "For filtering: A kind of value that the characteristic can take\
          \ on, such as numeric, text and so forth"
        required: false
        type: "string"
      - name: "serviceSpecCharacteristicConfigurable"
        in: "query"
        description: "For filtering: If true, the Boolean indicates that the ServiceSpecCharacteristic\
          \ is configurable"
        required: false
        type: "boolean"
      - name: "serviceSpecCharacteristicType"
        in: "query"
        description: "For filtering: (Class) type of the ServiceSpecCharacteristic"
        required: false
        type: "string"
      - name: "serviceSpecCharacteristicSchemaLocation"
        in: "query"
        description: "For filtering: A link to the schema describing this characteristic"
        required: false
        type: "string"
      - name: "serviceSpecCharacteristicValueSchemaLocation"
        in: "query"
        description: "For filtering: This (optional) field provides a link to the\
          \ schema describing the value type."
        required: false
        type: "string"
      - name: "serviceSpecCharacteristicMinCardinality"
        in: "query"
        description: "For filtering: The minimum number of instances a CharacteristicValue\
          \ can take on. For example, zero to five phone numbers in a group calling\
          \ plan, where zero is the value for the minCardinality."
        required: false
        type: "integer"
      - name: "serviceSpecCharacteristicMaxCardinality"
        in: "query"
        description: "For filtering: The maximum number of instances a CharacteristicValue\
          \ can take on. For example, zero to five phone numbers in a group calling\
          \ plan, where five is the value for the maxCardinality."
        required: false
        type: "integer"
      - name: "serviceSpecCharacteristicIsUnique"
        in: "query"
        description: "For filtering: An indicator that specifies if a value is unique\
          \ for the specification. Possible values are; \"unique while value is in\
          \ effect\" and \"unique whether value is in effect or not\""
        required: false
        type: "boolean"
      - name: "serviceSpecCharacteristicRegex"
        in: "query"
        description: "For filtering: A rule or principle represented in regular expression\
          \ used to derive the value of a characteristic value."
        required: false
        type: "string"
      - name: "serviceSpecCharacteristicExtensible"
        in: "query"
        description: "For filtering: An indicator that specifies that the values for\
          \ the characteristic can be extended by adding new values when instantiating\
          \ a characteristic for an Entity."
        required: false
        type: "boolean"
      - name: "relatedPartyRole"
        in: "query"
        description: "For filtering: Role of the related party."
        required: false
        type: "string"
      - name: "relatedPartyName"
        in: "query"
        description: "For filtering: Name of the related party"
        required: false
        type: "string"
      - name: "serviceSpecRelationshipType"
        in: "query"
        description: "For filtering: Type of relationship such as migration, substitution,\
          \ dependency, exclusivity"
        required: false
        type: "string"
      - name: "serviceSpecRelationshipRole"
        in: "query"
        description: "For filtering: The association role for this service specification"
        required: false
        type: "string"
      - name: "serviceSpecRelationshipName"
        in: "query"
        description: "For filtering: The name given to the target service specification\
          \ instance"
        required: false
        type: "string"
      - name: "targetServiceSchemaType"
        in: "query"
        description: "For filtering: Class type of the target service"
        required: false
        type: "string"
      - name: "targetServiceSchemaSchemaLocation"
        in: "query"
        description: "For filtering: This field provides a link to the schema describing\
          \ the target service"
        required: false
        type: "string"
      responses:
        200:
          description: "Ok"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/ServiceSpecification"
        400:
          description: "Bad Request"
          schema:
            $ref: "#/definitions/Error"
        500:
          description: "Internal Server Error"
          schema:
            $ref: "#/definitions/Error"
      x-swagger-router-controller: "ServiceSpecification"
    post:
      tags:
      - "serviceSpecification"
      summary: "Creates a 'ServiceSpecification'"
      operationId: "createServiceSpecification"
      parameters:
      - in: "body"
        name: "serviceSpecification"
        description: "The Service Specification to be created"
        required: true
        schema:
          $ref: "#/definitions/ServiceSpecification_Create"
      responses:
        201:
          description: "Created"
          schema:
            $ref: "#/definitions/ServiceSpecification"
        400:
          description: "Bad Request"
          schema:
            $ref: "#/definitions/Error"
        500:
          description: "Internal Server Error"
          schema:
            $ref: "#/definitions/Error"
      x-swagger-router-controller: "ServiceSpecification"
  /serviceSpecification/{id}:
    get:
      tags:
      - "serviceSpecification"
      summary: "Retrieves a 'ServiceSpecification' by Id"
      operationId: "retrieveServiceSpecification"
      parameters:
      - name: "id"
        in: "path"
        description: "Identifier of the Service Specification"
        required: true
        type: "string"
      responses:
        200:
          description: "Ok"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/ServiceSpecification"
        400:
          description: "Bad Request"
          schema:
            $ref: "#/definitions/Error"
        404:
          description: "Not Found"
          schema:
            $ref: "#/definitions/Error"
        500:
          description: "Internal Server Error"
          schema:
            $ref: "#/definitions/Error"
      x-swagger-router-controller: "ServiceSpecification"
    delete:
      tags:
      - "serviceSpecification"
      summary: "Deletes a 'ServiceSpecification' by Id"
      operationId: "deleteServiceSpecification"
      parameters:
      - name: "id"
        in: "path"
        description: "Identifier of the Service Specification"
        required: true
        type: "string"
      responses:
        204:
          description: "Deleted"
        400:
          description: "Bad Request"
          schema:
            $ref: "#/definitions/Error"
        404:
          description: "Not Found"
          schema:
            $ref: "#/definitions/Error"
        500:
          description: "Internal Server Error"
          schema:
            $ref: "#/definitions/Error"
      x-swagger-router-controller: "ServiceSpecification"
    patch:
      tags:
      - "serviceSpecification"
      summary: "Updates partially a 'ServiceSpecification' by Id"
      operationId: "patchServiceSpecification"
      parameters:
      - name: "id"
        in: "path"
        description: "Identifier of the Service Specification"
        required: true
        type: "string"
      - in: "body"
        name: "serviceSpecification"
        description: "The Service Specification to be updated"
        required: true
        schema:
          $ref: "#/definitions/ServiceSpecification_Update"
      responses:
        201:
          description: "Updated"
          schema:
            $ref: "#/definitions/ServiceSpecification"
        400:
          description: "Bad Request"
          schema:
            $ref: "#/definitions/Error"
        404:
          description: "Not Found"
          schema:
            $ref: "#/definitions/Error"
        500:
          description: "Internal Server Error"
          schema:
            $ref: "#/definitions/Error"
      x-swagger-router-controller: "ServiceSpecification"
...
definitions:
...
  ServiceSpecification:
    type: "object"
    properties:
      id:
        type: "string"
        description: "Unique identifier of this REST resource"
      href:
        type: "string"
        description: "Hyperlink reference to this REST resource"
      name:
        type: "string"
        description: "Name given to this REST resource"
      description:
        type: "string"
        description: "Description of this REST resource"
      '@type':
        type: "string"
        description: "Class type of this REST resource"
      '@schemaLocation':
        type: "string"
        description: "This field provides a link to the schema describing this REST\
          \ resource"
      '@baseType':
        type: "string"
        description: "The (immediate) base class type of this REST resource"
      version:
        type: "string"
        description: "Service specification version"
      validFor:
        description: "The period for which this REST resource is valid"
        $ref: "#/definitions/TimePeriod"
      lastUpdate:
        type: "string"
        format: "date-time"
        description: "Date and time of the last update of this REST resource"
      lifecycleStatus:
        type: "string"
        description: "Used to indicate the current lifecycle status of the service\
          \ specification"
      isBundle:
        type: "boolean"
        description: "A flag indicates that if this service specification is a bundled\
          \ specification (true) or single (false)."
      resourceSpecification:
        type: "array"
        items:
          $ref: "#/definitions/ResourceSpecificationRef"
      attachment:
        type: "array"
        items:
          $ref: "#/definitions/Attachment"
      serviceSpecCharacteristic:
        type: "array"
        items:
          $ref: "#/definitions/ServiceSpecCharacteristic"
      relatedParty:
        type: "array"
        items:
          $ref: "#/definitions/RelatedPartyRef"
      serviceSpecRelationship:
        type: "array"
        items:
          $ref: "#/definitions/ServiceSpecRelationship"
      targetServiceSchema:
        $ref: "#/definitions/TargetServiceSchemaRef"
    description: "ServiceSpecification is a class that offers characteristics to describe\
      \ a type of service.\nFunctionally, it acts as a template by which Services\
      \ may be instantiated. By sharing the same  specification, these services would\
      \ therefore share the same set of characteristics."
    example:
      isBundle: true
      targetServiceSchema:
        '@type': "@type"
        '@schemaLocation': "@schemaLocation"
      lifecycleStatus: "lifecycleStatus"
      validFor:
        startDateTime: "2000-01-23T04:56:07.000+00:00"
        endDateTime: "2000-01-23T04:56:07.000+00:00"
      '@type': "@type"
      resourceSpecification:
      - name: "name"
        id: "id"
        href: "href"
        version: "version"
      - name: "name"
        id: "id"
        href: "href"
        version: "version"
      description: "description"
      relatedParty:
      - role: "role"
        validFor:
          startDateTime: "2000-01-23T04:56:07.000+00:00"
          endDateTime: "2000-01-23T04:56:07.000+00:00"
        name: "name"
        id: "id"
        href: "href"
      - role: "role"
        validFor:
          startDateTime: "2000-01-23T04:56:07.000+00:00"
          endDateTime: "2000-01-23T04:56:07.000+00:00"
        name: "name"
        id: "id"
        href: "href"
      version: "version"
      '@baseType': "@baseType"
      attachment:
      - description: "description"
        href: "href"
        id: "id"
        type: "type"
        url: "url"
      - description: "description"
        href: "href"
        id: "id"
        type: "type"
        url: "url"
      lastUpdate: "2000-01-23T04:56:07.000+00:00"
      name: "name"
      id: "id"
      href: "href"
      '@schemaLocation': "@schemaLocation"
      serviceSpecCharacteristic:
      - validFor:
          startDateTime: "2000-01-23T04:56:07.000+00:00"
          endDateTime: "2000-01-23T04:56:07.000+00:00"
        '@type': "@type"
        '@valueSchemaLocation': "@valueSchemaLocation"
        isUnique: true
        description: "description"
        serviceSpecCharacteristicValue:
        - rangeInterval: "rangeInterval"
          isDefault: true
          valueTo: 5
          regex: "regex"
          unitOfMeasure: "unitOfMeasure"
          validFor:
            startDateTime: "2000-01-23T04:56:07.000+00:00"
            endDateTime: "2000-01-23T04:56:07.000+00:00"
          '@type': "@type"
          valueType: "valueType"
          '@schemaLocation': "@schemaLocation"
          value: "{}"
          valueFrom: 1
        - rangeInterval: "rangeInterval"
          isDefault: true
          valueTo: 5
          regex: "regex"
          unitOfMeasure: "unitOfMeasure"
          validFor:
            startDateTime: "2000-01-23T04:56:07.000+00:00"
            endDateTime: "2000-01-23T04:56:07.000+00:00"
          '@type': "@type"
          valueType: "valueType"
          '@schemaLocation': "@schemaLocation"
          value: "{}"
          valueFrom: 1
        maxCardinality: 6
        minCardinality: 0
        serviceSpecCharRelationship:
        - validFor:
            startDateTime: "2000-01-23T04:56:07.000+00:00"
            endDateTime: "2000-01-23T04:56:07.000+00:00"
          '@type': "@type"
          name: "name"
          id: "id"
          href: "href"
          type: "type"
        - validFor:
            startDateTime: "2000-01-23T04:56:07.000+00:00"
            endDateTime: "2000-01-23T04:56:07.000+00:00"
          '@type': "@type"
          name: "name"
          id: "id"
          href: "href"
          type: "type"
        regex: "regex"
        valueType: "valueType"
        name: "name"
        '@schemaLocation': "@schemaLocation"
        extensible: true
        configurable: true
      - validFor:
          startDateTime: "2000-01-23T04:56:07.000+00:00"
          endDateTime: "2000-01-23T04:56:07.000+00:00"
        '@type': "@type"
        '@valueSchemaLocation': "@valueSchemaLocation"
        isUnique: true
        description: "description"
        serviceSpecCharacteristicValue:
        - rangeInterval: "rangeInterval"
          isDefault: true
          valueTo: 5
          regex: "regex"
          unitOfMeasure: "unitOfMeasure"
          validFor:
            startDateTime: "2000-01-23T04:56:07.000+00:00"
            endDateTime: "2000-01-23T04:56:07.000+00:00"
          '@type': "@type"
          valueType: "valueType"
          '@schemaLocation': "@schemaLocation"
          value: "{}"
          valueFrom: 1
        - rangeInterval: "rangeInterval"
          isDefault: true
          valueTo: 5
          regex: "regex"
          unitOfMeasure: "unitOfMeasure"
          validFor:
            startDateTime: "2000-01-23T04:56:07.000+00:00"
            endDateTime: "2000-01-23T04:56:07.000+00:00"
          '@type': "@type"
          valueType: "valueType"
          '@schemaLocation': "@schemaLocation"
          value: "{}"
          valueFrom: 1
        maxCardinality: 6
        minCardinality: 0
        serviceSpecCharRelationship:
        - validFor:
            startDateTime: "2000-01-23T04:56:07.000+00:00"
            endDateTime: "2000-01-23T04:56:07.000+00:00"
          '@type': "@type"
          name: "name"
          id: "id"
          href: "href"
          type: "type"
        - validFor:
            startDateTime: "2000-01-23T04:56:07.000+00:00"
            endDateTime: "2000-01-23T04:56:07.000+00:00"
          '@type': "@type"
          name: "name"
          id: "id"
          href: "href"
          type: "type"
        regex: "regex"
        valueType: "valueType"
        name: "name"
        '@schemaLocation': "@schemaLocation"
        extensible: true
        configurable: true
      serviceSpecRelationship:
      - role: "role"
        validFor:
          startDateTime: "2000-01-23T04:56:07.000+00:00"
          endDateTime: "2000-01-23T04:56:07.000+00:00"
        name: "name"
        id: "id"
        href: "href"
        type: "type"
      - role: "role"
        validFor:
          startDateTime: "2000-01-23T04:56:07.000+00:00"
          endDateTime: "2000-01-23T04:56:07.000+00:00"
        name: "name"
        id: "id"
        href: "href"
        type: "type"
  ServiceSpecification_Create:
    type: "object"
    required:
    - "@type"
    - "name"
    properties:
      name:
        type: "string"
        description: "Name given to this REST resource"
      description:
        type: "string"
        description: "Description of this REST resource"
      '@type':
        type: "string"
        description: "Class type of this REST resource"
      '@schemaLocation':
        type: "string"
        description: "This field provides a link to the schema describing this REST\
          \ resource"
      '@baseType':
        type: "string"
        description: "The (immediate) base class type of this REST resource"
      version:
        type: "string"
        description: "Service specification version"
      validFor:
        description: "The period for which this REST resource is valid"
        $ref: "#/definitions/TimePeriod"
      lastUpdate:
        type: "string"
        format: "date-time"
        description: "Date and time of the last update of this REST resource"
      lifecycleStatus:
        type: "string"
        description: "Used to indicate the current lifecycle status of the service\
          \ specification"
      isBundle:
        type: "boolean"
        description: "A flag indicates that if this service specification is a bundled\
          \ specification (true) or single (false)."
      resourceSpecification:
        type: "array"
        items:
          $ref: "#/definitions/ResourceSpecificationRef"
      attachment:
        type: "array"
        items:
          $ref: "#/definitions/Attachment"
      serviceSpecCharacteristic:
        type: "array"
        items:
          $ref: "#/definitions/ServiceSpecCharacteristic"
      relatedParty:
        type: "array"
        items:
          $ref: "#/definitions/RelatedPartyRef"
      serviceSpecRelationship:
        type: "array"
        items:
          $ref: "#/definitions/ServiceSpecRelationship"
      targetServiceSchema:
        $ref: "#/definitions/TargetServiceSchemaRef"
    description: "ServiceSpecification is a class that offers characteristics to describe\
      \ a type of service.\nFunctionally, it acts as a template by which Services\
      \ may be instantiated. By sharing the same  specification, these services would\
      \ therefore share the same set of characteristics.\nSkipped properties: id,href"
    example:
      isBundle: true
      targetServiceSchema:
        '@type': "@type"
        '@schemaLocation': "@schemaLocation"
      lifecycleStatus: "lifecycleStatus"
      validFor:
        startDateTime: "2000-01-23T04:56:07.000+00:00"
        endDateTime: "2000-01-23T04:56:07.000+00:00"
      '@type': "@type"
      resourceSpecification:
      - name: "name"
        id: "id"
        href: "href"
        version: "version"
      - name: "name"
        id: "id"
        href: "href"
        version: "version"
      description: "description"
      relatedParty:
      - role: "role"
        validFor:
          startDateTime: "2000-01-23T04:56:07.000+00:00"
          endDateTime: "2000-01-23T04:56:07.000+00:00"
        name: "name"
        id: "id"
        href: "href"
      - role: "role"
        validFor:
          startDateTime: "2000-01-23T04:56:07.000+00:00"
          endDateTime: "2000-01-23T04:56:07.000+00:00"
        name: "name"
        id: "id"
        href: "href"
      version: "version"
      '@baseType': "@baseType"
      attachment:
      - description: "description"
        href: "href"
        id: "id"
        type: "type"
        url: "url"
      - description: "description"
        href: "href"
        id: "id"
        type: "type"
        url: "url"
      lastUpdate: "2000-01-23T04:56:07.000+00:00"
      name: "name"
      '@schemaLocation': "@schemaLocation"
      serviceSpecCharacteristic:
      - validFor:
          startDateTime: "2000-01-23T04:56:07.000+00:00"
          endDateTime: "2000-01-23T04:56:07.000+00:00"
        '@type': "@type"
        '@valueSchemaLocation': "@valueSchemaLocation"
        isUnique: true
        description: "description"
        serviceSpecCharacteristicValue:
        - rangeInterval: "rangeInterval"
          isDefault: true
          valueTo: 5
          regex: "regex"
          unitOfMeasure: "unitOfMeasure"
          validFor:
            startDateTime: "2000-01-23T04:56:07.000+00:00"
            endDateTime: "2000-01-23T04:56:07.000+00:00"
          '@type': "@type"
          valueType: "valueType"
          '@schemaLocation': "@schemaLocation"
          value: "{}"
          valueFrom: 1
        - rangeInterval: "rangeInterval"
          isDefault: true
          valueTo: 5
          regex: "regex"
          unitOfMeasure: "unitOfMeasure"
          validFor:
            startDateTime: "2000-01-23T04:56:07.000+00:00"
            endDateTime: "2000-01-23T04:56:07.000+00:00"
          '@type': "@type"
          valueType: "valueType"
          '@schemaLocation': "@schemaLocation"
          value: "{}"
          valueFrom: 1
        maxCardinality: 6
        minCardinality: 0
        serviceSpecCharRelationship:
        - validFor:
            startDateTime: "2000-01-23T04:56:07.000+00:00"
            endDateTime: "2000-01-23T04:56:07.000+00:00"
          '@type': "@type"
          name: "name"
          id: "id"
          href: "href"
          type: "type"
        - validFor:
            startDateTime: "2000-01-23T04:56:07.000+00:00"
            endDateTime: "2000-01-23T04:56:07.000+00:00"
          '@type': "@type"
          name: "name"
          id: "id"
          href: "href"
          type: "type"
        regex: "regex"
        valueType: "valueType"
        name: "name"
        '@schemaLocation': "@schemaLocation"
        extensible: true
        configurable: true
      - validFor:
          startDateTime: "2000-01-23T04:56:07.000+00:00"
          endDateTime: "2000-01-23T04:56:07.000+00:00"
        '@type': "@type"
        '@valueSchemaLocation': "@valueSchemaLocation"
        isUnique: true
        description: "description"
        serviceSpecCharacteristicValue:
        - rangeInterval: "rangeInterval"
          isDefault: true
          valueTo: 5
          regex: "regex"
          unitOfMeasure: "unitOfMeasure"
          validFor:
            startDateTime: "2000-01-23T04:56:07.000+00:00"
            endDateTime: "2000-01-23T04:56:07.000+00:00"
          '@type': "@type"
          valueType: "valueType"
          '@schemaLocation': "@schemaLocation"
          value: "{}"
          valueFrom: 1
        - rangeInterval: "rangeInterval"
          isDefault: true
          valueTo: 5
          regex: "regex"
          unitOfMeasure: "unitOfMeasure"
          validFor:
            startDateTime: "2000-01-23T04:56:07.000+00:00"
            endDateTime: "2000-01-23T04:56:07.000+00:00"
          '@type': "@type"
          valueType: "valueType"
          '@schemaLocation': "@schemaLocation"
          value: "{}"
          valueFrom: 1
        maxCardinality: 6
        minCardinality: 0
        serviceSpecCharRelationship:
        - validFor:
            startDateTime: "2000-01-23T04:56:07.000+00:00"
            endDateTime: "2000-01-23T04:56:07.000+00:00"
          '@type': "@type"
          name: "name"
          id: "id"
          href: "href"
          type: "type"
        - validFor:
            startDateTime: "2000-01-23T04:56:07.000+00:00"
            endDateTime: "2000-01-23T04:56:07.000+00:00"
          '@type': "@type"
          name: "name"
          id: "id"
          href: "href"
          type: "type"
        regex: "regex"
        valueType: "valueType"
        name: "name"
        '@schemaLocation': "@schemaLocation"
        extensible: true
        configurable: true
      serviceSpecRelationship:
      - role: "role"
        validFor:
          startDateTime: "2000-01-23T04:56:07.000+00:00"
          endDateTime: "2000-01-23T04:56:07.000+00:00"
        name: "name"
        id: "id"
        href: "href"
        type: "type"
      - role: "role"
        validFor:
          startDateTime: "2000-01-23T04:56:07.000+00:00"
          endDateTime: "2000-01-23T04:56:07.000+00:00"
        name: "name"
        id: "id"
        href: "href"
        type: "type"
...
  Error:
    required:
    - "code"
    - "message"
    properties:
      code:
        type: "integer"
        description: "An integer coding the error type. This is given to caller so\
          \ he can translate them if required."
      message:
        type: "string"
        description: "A short localized string that describes the error."
      description:
        type: "string"
        description: "(optional) A long localized error description if needed. It\
          \ can contain precise information about which parameter is missing, or what\
          \ are the identifier acceptable values."
      infoURL:
        type: "string"
        description: "(optional) A URL to online documentation that provides more\
          \ information about the error."